Collapsable_viz. Please read the comments.
This is the HTML file that does the visualisation. The data file referenced here has to be prepared separately. It will not work without the data file and I have not uploaded it yet until I finish the preprocessing today. Maybe we might need to discard zero values. And again when I finish preparing this we need to set it up on a local server before it works. Please don't be disappointed if you clean on it and it doesn't work! It needs data file and local server setup ... The local server setup requirement  can be bypassed with google extension "200 ok!"